添加监测状态 开始准备发送的数据

+++ app/src/main/java/com/puyu/mobile/sdi/bean/SystemMonitor.java
@@ -10,7 +10,7 @@ package com.puyu.mobile.sdi.bean;
 public class SystemMonitor {
     public SysStateEnum sysState;//系统状态
     public byte runProcess;//当前运行流程
-    public int runPassage;//当前运行通道数
+    public byte runPassage;//当前运行通道数
     public float currentPress;//当前压力
     public float targetPress;//目标压力
     public float averagePress;//平均压力
